pyproject.toml
[project]
name = "gravl"
version = "0.1.0"
description = "Integration & unification hub — Shopify / Cashfree / Freshdesk / Eshopbox / WhatsApp"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.1",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "gspread>=6.1",
    "google-auth>=2.30",
    "prefect>=3.0",
    "tenacity>=8.4",
    "structlog>=24.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gravl"]

[tool.ruff]
line-length = 100
target-version = "py312"
